[NPT-9220] Move test_delete_artifact test from experiment to backend
